\name{toprab}
\alias{toprab}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Convert abundance matrix into binary matrix}
\description{
  Converts abundance matrix into binary presence/absence matrix (1 if
  abundance>0). 
}
\usage{
toprab(prabobj)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{prabobj}{object of class \code{prab}.}
}
\value{
  matrix generated by \code{prabobj$prab>0}.
}

\author{Christian Hennig
  \email{chrish@stats.ucl.ac.uk}
  \url{http://www.homepages.ucl.ac.uk/~ucakche}}


\examples{
  data(siskiyou)
  x <- prabinit(prabmatrix=siskiyou, neighborhood=siskiyou.nb,
             distance="none")
  toprab(x)
}
\keyword{manip}% at least one, from doc/KEYWORDS

