\name{waterdist}
\alias{waterdist}
% \non_function{}
\title{Overwater distances between islands in the Aegean sea}
\description{
  Distance matrix of overwater distances in km between 34 islands in the
  Aegean sea.
}
\usage{data(waterdist)}
\format{
  A symmetric 34*34 distance matrix.}
\source{
  B. Hausdorf and C. Hennig (2005) The influence of recent geography,
  palaeography and climate on the composition of the faune of the
  central Aegean Islands.   \emph{Biological Journal of the Linnean
    Society} 84, 785-795.
}
\details{
  Reads from example data file \code{Waterdist.dat}, in which there is a
  35th column and line with distances to Turkey mainland.
  }

\examples{
data(waterdist)
}
\keyword{datasets}
% \keyword{spatial}

